+++ client/pages/home/index.js
@@ -1,29 +1,30 @@
 import React from 'react';
 import {render} from 'react-dom';
 import {combineReducers} from 'redux-immutable';
-import {Route, IndexRoute} from 'react-router';
+import {Route} from 'react-router-dom';
 
 import Root from '../../Root';
-import routing from '../../common/reducers/routing';
 import toast from '../../common/reducers/toast';
 import home from './reducer';
 import App from '../../common/App';
 import HomePage from './HomePage';
 import {urlContext} from '../../utils/config';
 
-const routes = (
-  <Route path="/" component={App}>
-    <IndexRoute component={HomePage}/>
-  </Route>
-);
+const routes = (store) => {
+  return (
+    <App>
+      <Route path="/" component={HomePage}/>
+    </App>
+  );
+};
+
 
 const reducers = combineReducers({
-  routing,
   toast,
-  home
+  home,
 });
 
 render(
   <Root routes={routes} reducers={reducers} basename={`${urlContext}/home`}/>,
-  document.getElementById('layout')
+  document.getElementById('layout'),
 );