+++ client/pages/page-2/index.js
@@ -3,7 +3,6 @@ import {render} from 'react-dom';
 import {combineReducers} from 'redux-immutable';
 
 import Root from '../../Root';
-import routing from '../../common/reducers/routing';
 import toast from '../../common/reducers/toast';
 import caches from '../../common/reducers/caches';
 import person from './person/reducer';
@@ -13,7 +12,6 @@ import routes from './routes';
 import {urlContext} from '../../utils/config';
 
 const reducers = combineReducers({
-  routing,
   toast,
   caches,
   person,