+++ client/pages/page-2/routes.js
@@ -1,51 +1,30 @@
 import React from 'react';
-import {Route, IndexRoute} from 'react-router';
+import {Route, Switch} from 'react-router-dom';
+import App from '../../common/App';
+import Page from './Page';
 import {enterHandler, leaveHandler} from './route-setting';
+import Person from './person/components/Person';
+import PersonList from './person/components/PersonList';
+import PersonForm from './person/components/PersonForm';
+import Film from './film/components/Film';
 
 // 把 store 传入路由中，这样不同的路由可以根据 store 值来处理业务逻辑
-/*eslint-disable react/jsx-no-bind,react/react-in-jsx-scope*/
+/*eslint-disable react/jsx-no-bind,react/react-in-jsx-scope,react/no-children-prop*/
 const routes = (store) => {
   return (
-    <Route path="/" getComponent={(nextState, cb) => {
-      require.ensure([], require => {
-        cb(null, require('../../common/App').default);
-      });
-    }}>
-      <Route path="/" getComponent={(nextState, cb) => {
-        require.ensure([], require => {
-          cb(null, require('./Page').default);
-        });
-      }}>
-        <Route path="/person" onEnter={enterHandler('person')}
-               onLeave={leaveHandler('person')}
-               getComponent={(nextState, cb) => {
-          require.ensure([], require => {
-            cb(null, require('./person/components/Person').default);
-          });
+    <App>
+      <Page>
+        <Route path="/" children={(props) => {
+          console.info(props);
+          return (
+            <li>
+              ddd
+            </li>
+          );
         }}>
-          <IndexRoute getComponent={(nextState, cb) => {
-            require.ensure([], require => {
-              cb(null, require('./person/components/PersonList').default);
-            });
-          }}/>
-          <Route path="create" onEnter={enterHandler('create')}
-                 onLeave={leaveHandler('create')}
-                 getComponent={(nextState, cb) => {
-            require.ensure([], require => {
-              cb(null, require('./person/components/PersonForm').default);
-            });
-          }}/>
         </Route>
-        <Route path="/film" onEnter={enterHandler('film')}
-               onLeave={leaveHandler('film')}
-               getComponent={(nextState, cb) => {
-          require.ensure([], require => {
-            cb(null, require('./film/components/Film').default);
-          });
-        }}/>
-      </Route>
-    </Route>
+      </Page>
+    </App>
   );
 };
-
 export default routes;