+++ scripts/webpack.config.dll.babel.js
@@ -1,15 +1,16 @@
 import path from 'path';
 import webpack from 'webpack';
 
-const appRoot = path.join(__dirname,'../');
+const appRoot = path.join(__dirname, '../');
 
 export default {
+  mode: 'development',
   entry: {
-    vendor: ['classnames', 'history/lib/createBrowserHistory', 'immutable',
+    vendor: ['classnames', 'eruda', 'history', 'immutable',
       'isomorphic-fetch', 'prop-types', 'react', 'react-dom',
-      'react-redux', 'react-router', 'react-router-redux-fixed', 'react-transition-group',
+      'react-redux', 'react-router', 'react-router-dom', 'react-router-redux', 'react-transition-group',
       'redux', 'redux-immutable', 'redux-thunk', 'redux-devtools',
-      'redux-devtools-dock-monitor', 'redux-devtools-log-monitor', 'redux-logger']
+      'redux-devtools-dock-monitor', 'redux-devtools-log-monitor', 'redux-logger'],
   },
 
   output: {
@@ -20,13 +21,13 @@ export default {
      * 将会定义为 window.${output.library}
      * 在这次的例子中，将会定义为`window.vendorLibrary`
      */
-    library: '[name]Library'
+    library: '[name]Library',
   },
   plugins: [
     new webpack.DefinePlugin({
       'process.env': {
         NODE_ENV: JSON.stringify('development'),
-      }
+      },
     }),
     new webpack.DllPlugin({
       /**
@@ -40,7 +41,7 @@ export default {
        * dll bundle 输出到那个全局变量上
        * 和 output.library 一样即可。
        */
-      name: '[name]Library'
-    })
-  ]
+      name: '[name]Library',
+    }),
+  ],
 };