+++ scripts/webpack.config.prod.babel.js
@@ -4,10 +4,9 @@ import ExtractTextPlugin from 'extract-text-webpack-plugin';
 import ManifestPlugin from 'webpack-manifest-plugin';
 import autoprefixer from 'autoprefixer';
 import flexbugs from 'postcss-flexbugs-fixes'; // 修复 flexbox 已知的 bug
-import Webpack2Polyfill from 'webpack2-polyfill-plugin';
-
-const cssnano = require('cssnano'); // 优化 css，对于长格式优化成短格式等
+import cssnano from 'cssnano'; // 优化 css，对于长格式优化成短格式等
 import incstr from 'incstr';
+
 // 根目录上下文
 import { urlContext } from '../client/utils/config';
 
@@ -117,15 +116,14 @@ const extractCSS = new ExtractTextPlugin({
 
 // 基于 webpack 的持久化缓存方案 可以参考 https://github.com/pigcan/blog/issues/9
 const webpackConfig = {
+  mode: 'production',
   devtool: 'source-map', // 生成 source-map文件 原始源码
   target: 'web', // webpack 能够为多种环境构建编译, 默认是 'web'，可省略 https://doc.webpack-china.org/configuration/target/
   resolve: {
     // 自动扩展文件后缀名
     extensions: ['.js', '.scss', '.css', '.png', '.jpg', '.gif'],
     // 模块别名定义，方便直接引用别名
-    alias: {
-      'react-router-redux': path.resolve(nodeModules, 'react-router-redux-fixed/lib/index.js'),
-    },
+    alias: {},
     // 参与编译的文件
     modules: [
       'client',
@@ -164,19 +162,23 @@ const webpackConfig = {
           options: {
             babelrc: false,
             cacheDirectory: true,
+            // babel-preset-env 的配置可参考 https://zhuanlan.zhihu.com/p/29506685
+            // 他会自动使用插件和 polyfill
             presets: [
-              'react', 'stage-3', ['env', {
-                modules: false,
+              'react', ['env', {
+                modules: false, // 设为 false，交由 Webpack 来处理模块化
                 targets: {
                   browsers
                 },
+                // 设为 true 会根据需要自动导入用到的 es6 新方法，而不是一次性的引入 babel-polyfill
+                // 比如使用 Promise 会导入 import "babel-polyfill/core-js/modules/es6.promise";
                 useBuiltIns: true,
               }],
             ],
             plugins: [
-              'syntax-dynamic-import', //支持'import()'
-              'transform-class-properties', //解析类属性，静态和实例的属性
-              'transform-object-assign', //polyfill object-assign
+              'syntax-dynamic-import', // 支持'import()'
+              'transform-class-properties', // 解析类属性，静态和实例的属性
+              'transform-object-rest-spread', // 支持对象 rest
               [
                 'transform-react-remove-prop-types',
                 {
@@ -197,7 +199,7 @@ const webpackConfig = {
           loader: 'url-loader',
           options: {
             name: '[hash:8].[ext]',
-            limit: 8192, // 8kb
+            limit: 10000, // 10kb
           },
         },
       },
@@ -253,17 +255,31 @@ const webpackConfig = {
     ],
   },
 
+  // webpack 选项配置
+  optimization: {
+    noEmitOnErrors: true, // 在编译出现错误时，用来跳过输出阶段
+    occurrenceOrder: true, // Webpack将会用更短的名字去命名引用频度更高的chunk
+    sideEffects: false,
+    minimize: true,
+    concatenateModules: true, // Scope Hoisting-作用域提升
+    splitChunks: {
+      cacheGroups: {
+        vendor: {
+          // test: /node_modules/, // 指定文件夹
+          name: 'vendor',
+          chunks: 'all'
+        }
+      }
+    },
+    runtimeChunk: {
+      name: 'manifest'
+    },
+    namedChunks: true,
+  },
+
   plugins: [
-    new Webpack2Polyfill(),
-    new webpack.NoEmitOnErrorsPlugin(),
     // 用来优化生成的代码 chunk，合并相同的代码
     new webpack.optimize.AggressiveMergingPlugin(),
-    new webpack.optimize.ModuleConcatenationPlugin({// Scope Hoisting-作用域提升
-      // 检查所有的模块
-      maxModules: Infinity,
-      // 将显示绑定失败的原因
-      optimizationBailout: true,
-    }),
     new webpack.DefinePlugin({
       'process.env': {
         NODE_ENV: JSON.stringify('production'),
@@ -271,41 +287,19 @@ const webpackConfig = {
     }),
     extractScss,
     extractCSS,
-    // 以下两个插件可以解决持久化缓存，但由于用到了模块化，混淆的 css 名导致无法缓存
-    // 故先注释掉，待后续解决了，再打开，或者不使用 css Module 时
     new webpack.HashedModuleIdsPlugin(),
-    new webpack.NamedChunksPlugin(),
-    // https://doc.webpack-china.org/guides/code-splitting-libraries/#manifest-
-    new webpack.optimize.CommonsChunkPlugin('vendor'),
     new ManifestPlugin({
       basePath: `${urlContext}/dist/`,
-    }),
-    new webpack.optimize.UglifyJsPlugin({
-      sourceMap: true,
-      compressor: {
-        warnings: false,
-        /*eslint-disable camelcase*/
-        drop_console: process.env.NODE_ENV === 'production' // 只有正式环境去掉 console
-      },
-      mangle: {
-        except: [] // 设置不混淆变量名
-      },
     }),
     new webpack.BannerPlugin({
       banner: [
         '/*!',
         ' react-redux-router-base',
-        ` Copyright © 2017-${new Date().getFullYear()} JD Finance.`,
+        ` Copyright © 2018-${new Date().getFullYear()} JD Finance.`,
         '*/',
       ].join('\n'),
       raw: true,
       entryOnly: true,
-    }),
-    new webpack.LoaderOptionsPlugin({
-      /*UglifyJsPlugin 不再压缩 loaders。在未来很长一段时间里，需要通过设置 minimize:true 来压缩 loaders。
-       loaders 的压缩模式将在 webpack 3 或后续版本中取消。
-       为了兼容旧的 loaders，loaders 可以通过插件来切换到压缩模式：*/
-      minimize: true,
     }),
   ],
 };