+++ server/index.js
@@ -46,7 +46,7 @@ if (isDev) {
     cache: true, // 开启缓存，增量编译
     stats: {
       colors: true, // 打印日志显示颜色
-      reasons: true // 打印相关被引入的模块
+      // reasons: true // 打印相关被引入的模块
     },
     publicPath: webpackConfig.output.publicPath
   }));